Enhances task and list management security

+++ src/main/java/com/todoapp/project/adapter/out/ProjectRepositoryImpl.java
@@ -4,8 +4,6 @@ import com.todoapp.project.application.mapper.ProjectMapper;
 import com.todoapp.project.domain.Project;
 import com.todoapp.project.port.out.ProjectRepository;
 import com.todoapp.user.adapter.out.UserEntity;
-import com.todoapp.user.adapter.out.UserMapper;
-import com.todoapp.user.port.out.UserRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.stereotype.Repository;
@@ -49,11 +47,23 @@ public class ProjectRepositoryImpl implements ProjectRepository {
     @Transactional(readOnly = true)
     @Override
     public List<Project> findByUserId(UUID userId) {
-        return mapper.entitiesToDomains(jpaRepository.findByOwnerId(userId));
+        List<ProjectEntity> entities = jpaRepository.findByOwnerId(userId);
+        if (entities.isEmpty()) {
+            throw new NoSuchElementException("No se encontraron proyectos para el usuario con id: " + userId);
+        }
+        return mapper.entitiesToDomains(entities);
     }
 
     @Override
     public void delete(UUID id) {
+    if (!jpaRepository.existsById(id)) {
+            throw new NoSuchElementException("No se encontró el proyecto con id: " + id);
+    }
         jpaRepository.deleteById(id);
+    }
+
+    @Override
+    public boolean existsById(UUID id) {
+        return jpaRepository.existsById(id);
     }
 }
